==> 2_humancohort_unsupervisedStats.Rmd <==
---
title: "ICC - Mixomics whole cohort"
author: "Miguel Cosenza & Tilman Werner - AG Schilling - Uniklinik Freiburg"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document: null
  toc: yes
  toc_float: yes
  pdf_document: default
---

```{r setup, include=FALSE}
rm(list = ls())
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width=5, fig.height=4, error=TRUE)
```

```{r message=FALSE, warning=FALSE, echo=FALSE}
## Required packages ----
library(tidyverse)
library(readxl)
library(mixOmics)
library(kableExtra)
library(reactable)
library(proDA)
library(missForest)
library(dplyr)
library(tidyr)
library(ComplexHeatmap)
library(circlize)
library(ggpubr)
library(grid)
library(gtsummary)

`%nin%` = Negate(`%in%`)
```

## Load data and general wrangling  

comment: 

KB151 was mistakenly labeled as KB154 in batch2. The names of the files etc. were corrected accordingly.
```{r}
## Load and prepare data and annotation ----
oridata <- read.csv("results_1_humancohort_LoadData/TimsTOF_ICC_humancohort_matrix.tsv") %>%
  dplyr::rename(Protein = "X")

colnames_data <- colnames(oridata) %>%
  str_replace("KB154_S3.G5_1_11631.d", "KB151_S3.G5_1_11631.d") %>% 
  str_remove("(?<=_).+") %>%
  str_remove("_") %>%
  str_remove("B$")

colnames(oridata) <- colnames_data

annotation <- read_excel("Data/Table S1 - Patient Table.xlsx") %>%
              filter(Sample %nin% c("KB106", "KB42", "KB147", "KB53", "KB213", "KB215", "KB55", "KB23", "KB209")) %>% #"KB107", "KB167", 
              arrange(Sample)

oridata1 <- dplyr::select(oridata, one_of("Protein", annotation$Sample)) %>%
  mutate_if(is.numeric, function(x, na.rm = TRUE) log2(x)) %>%
  mutate_if(is.numeric, function(x, na.rm = TRUE) x - median(x, na.rm = TRUE))

# Load batch correction info
annot_batch <- read_excel("Data/annotation_batch_correction.xlsx") %>%
  dplyr::rename(Sample = "FullRunName")

annotation <- left_join(annotation, annot_batch, by = "Sample")

oridata1 <- dplyr::select(oridata1, "Protein", one_of(annotation$Sample))
annotation <- filter(annotation, Sample %in% colnames(oridata1))

annotation1 <- dplyr::select(annotation, Sample, Group)
predictors1 <- annotation$Group

write.csv(oridata1, file = "results_2_humancohort_unsupervisedStats/TimsTOF_ICC_ExprMat_log2_median.csv")
write.csv(annotation, file = "results_2_humancohort_unsupervisedStats/TimsTOF_ICC_FullAnnotation.csv")

# create Patient Table 1
summary1 <- filter(annotation, Group == "Tumor") %>%
  dplyr::select("Age (years)", "Ethnicity", "Sex", "PSC", "Hep B/C",
                          "Tumor Stage", "Radiotherapy", "Duct Type", "Lymphangioinvasion", "Satellite Nodules", 
                          "Liver Capsule Invasion", "Portal Spread", 
                          "Overall Tumor Budding", "intratumoral Tumor Budding", "peritumoral Tumor Budding") 

summary1[summary1 == "NA"] <- NA
summary1[summary1 == "NT"] <- NA
summary1[summary1 == "NR"] <- NA
summary1[summary1 == "NotTested"] <- NA
summary1 <- droplevels(summary1)

gtsumm <- tbl_summary(summary1) %>%
  modify_header(label = "**ICC cohort**")
gtsumm
as_gt(gtsumm) %>%
  gt::gtsave(filename = "results_2_humancohort_unsupervisedStats/ICC_PatTable.html")
```

Plot for ID numbers
```{r}
graph <- data.frame(colnames(oridata1[,-1]), colSums(!is.na(oridata1[,-1])), 1:ncol(oridata1[,-1])) 
colnames(graph) <- c("Sample", "IDnumber", "Samples") 

ggplot(graph, aes(Samples, sort(IDnumber))) +
  geom_point(size = 1) +
  labs(title = "Identified Protein Groups per Sample", y = "Protein Groups") +
  scale_y_continuous(limits = c(0, 9000))

plot1 <- ggplot(graph, aes(x = Samples, y = sort(IDnumber))) +
  geom_col(fill = "blanchedalmond", color = "black") +
  labs(y = "Identified Protein Groups") +
  scale_y_continuous(limits = c(0, 9000)) +
  theme_minimal() 
  # theme(axis.text=element_text(size=12),
  #       axis.title=element_text(size=14,face="bold"))
plot1

tiff("results_2_humancohort_unsupervisedStats/IDnumbers.tiff", units = "in", width = 5, height = 4, res = 300, pointsize =(5*100/72))
plot1
dev.off()

graph <- left_join(graph, annotation1, by = "Sample")

plot1B <- ggplot(graph, aes(x = Samples, y = sort(IDnumber), fill = Group)) +
  geom_col(color = "black") +
  labs(y = "Identified Protein Groups") +
  scale_y_continuous(limits = c(0, 7000)) +
  theme_minimal() 
  # theme(axis.text=element_text(size=12),
  #       axis.title=element_text(size=14,face="bold"))
plot1B

tiff("results_2_humancohort_unsupervisedStats/IDnumbers_Tissue.tiff", units = "in", width = 5, height = 4, res = 300, pointsize =(5*100/72))
plot1B
dev.off()

paste("average identified protein groups = ", mean(graph$IDnumber))

# plot ID numbers relative to measurement batch and order

graph_batch <- left_join(graph, annot_batch, by = "Sample") %>%
  mutate(MS_batch = as.character(MS_batch))

ggplot(graph_batch, aes(x = MS_batch, y = IDnumber)) +
  geom_boxplot() +
  geom_jitter() +
  theme_minimal()

plot4 <- ggplot(graph_batch, aes(x = reorder(Samples, measurement_order), y = IDnumber, color = Group)) +
  geom_point(size = 1) +
  labs(title = "Identified Protein Groups per Sample", y = "Protein Groups") +
  scale_y_continuous(limits = c(0, 9000)) +
  scale_color_manual(breaks = c("normal", "Tumor"),
                     values = c("cadetblue", "red")) +
  theme_classic()
plot4

tiff("results_2_humancohort_unsupervisedStats/IDs_time.tiff", units = "in", width = 6, height = 4, res = 300, pointsize =(6*100/72))
plot4
dev.off()

plot5 <- ggplot(graph_batch, aes(x = Group, y = IDnumber, fill = Group)) +
  geom_violin() +
  geom_boxplot(width = 0.3) +
  scale_fill_manual(breaks = c("normal", "Tumor"),
                     values = c("#4682B4", "#A52A2A")) +
  geom_pwc(method = "t.test", label = "p.format", hide.ns = TRUE) +
  #labs(caption = "two-sided t.test") +
  theme_classic() +
  theme(legend.position = "none") +
  ylab("ID count") +
  xlab("")
plot5

ggsave("results_2_humancohort_unsupervisedStats/IDs_Tissue.pdf", plot = plot5, width = 50, height = 46, 
       units = "mm", dpi = 300, scale = 2)

svg("results_2_humancohort_unsupervisedStats/IDs_Tissue.svg", width = 1.96, height = 1.81) #, pointsize =(2*100/72)
plot5
dev.off()
```

Plot for missingness across cohort
```{r}
expr_dat <- mutate(oridata1, sumNA = rowSums(is.na(oridata1)))  %>%
            mutate(completeness = (1 - (sumNA / (ncol(oridata1)-1))))

graph   <-  dplyr::select(expr_dat, c("Protein", "sumNA", "completeness")) %>%
            arrange(-completeness) %>%
            mutate(ProteinGroups = 1:nrow(expr_dat))

lessthan80percent <- filter(graph, completeness >= 0.8)

graph$completeness <- (1-graph$completeness) * 100

plot2 <- ggplot(data = graph, aes(ProteinGroups, completeness)) +
  geom_line() +
    labs(subtitle = paste(nrow(lessthan80percent), " Proteins with more than 80% coverage"),
       x = "Identified Protein Groups", y= "Missing Values [%]") +
  geom_hline(yintercept = 20, linetype = "dashed", color = "red") +
  theme_classic() +
  scale_y_reverse()
plot2

ggsave("results_2_humancohort_unsupervisedStats/NAs.pdf", plot = plot2, width = 63, height = 42, 
       units = "mm", dpi = 300, scale = 2)

tiff("results_2_humancohort_unsupervisedStats/NAs.tiff", units = "in", width = 5, height = 4, res = 300, pointsize =(5*100/72))
plot2
dev.off()
```

# Data pre-processing  
For unsupervised statistics (clustering, PCA) an imputed data matrix is used, since these methods cannot process missing values. All other analyses rely on the original, unimputed matrix.
```{r}
t_data_1st_log2_med_normaft <- column_to_rownames(oridata1, var = "Protein") %>%
  filter(rowSums(!is.na(.)) / ncol(.) >= 0.8) %>%
  t()

### Random forest missing value imputation ----

if(!file.exists("results_2_humancohort_unsupervisedStats/imputfr_aft.rds")){
  set.seed(546)
  #tictoc::tic()
  impfr_2 <- missForest(t_data_1st_log2_med_normaft) # this takes ~ 3 minutes to compute
  #tictoc::toc()
  write_rds(impfr_2, file = "results_2_humancohort_unsupervisedStats/imputfr_aft.rds")
} else {
  impfr_2 <- read_rds(file = "results_2_humancohort_unsupervisedStats/imputfr_aft.rds")
}

t_data_1st_log2_med_normimpaft <- impfr_2$ximp

data_1st_log2_med_normimpaft <- t(t_data_1st_log2_med_normimpaft)
```   

# Principal component analysis (PCA)  
```{r}
pca_res = mixOmics::pca(t_data_1st_log2_med_normimpaft, ncomp = 10, center = TRUE, scale = TRUE)

plot(pca_res)

pca <- plotIndiv(pca_res, ind.names = FALSE)
pca_df <- pca$df %>%
  rownames_to_column(var = "Sample") %>%
  mutate(Sample = str_remove_all(Sample, "Tumor-")) %>%
  inner_join(annotation, by = "Sample") %>%
  dplyr::rename(Tissue = "Group")

plot3 <- ggplot(pca_df, aes(x, y, color = Tissue)) +
  geom_point(size = 2) +
  theme_classic() +
  theme(legend.position = "none") +
  xlab(pca$graph$labels$x) +
  ylab(pca$graph$labels$y) +
  scale_color_manual(values = c("#4682B4", "#A52A2A"),
                     breaks = c("normal", "Tumor"))
plot3

ggsave("results_2_humancohort_unsupervisedStats/PCA_Tu-TANM.pdf", plot = plot3, width = 61, height = 46, 
       units = "mm", dpi = 300, scale = 2)

tiff("results_2_humancohort_unsupervisedStats/PCA_Tu-TANM.tiff", units = "in", width = 5, height = 6, res = 300, pointsize =(6*100/72))
plot3
dev.off()
```   

# Hierarchical Clustering   
```{r, fig.width=12, fig.height=6, results="asis"}
my_dist_t <- dist(t_data_1st_log2_med_normaft)
my_hclust_t <- hclust(my_dist_t)
plot(my_hclust_t)
```

#Heatmap
Indicating measurement and digestion batches
```{r}
oridata2 <- column_to_rownames(oridata1, var = "Protein")
oridata2 <- filter(oridata2, rowSums(!is.na(oridata2)) / ncol(oridata2) > 0.8) %>%
  dplyr::select(one_of(annotation$Sample)) 
subtractor <- apply(oridata2[,-1], 1, median, na.rm = TRUE)
oridata2 <- oridata2 - subtractor
oridata2 <- as.matrix(oridata2) 

annotation$Group <- str_replace_all(annotation$Group, "normal", "TANM")

#CAVE! annotation file and plot matrix need to be in the exactly same order to match in the plot!
ha = columnAnnotation(Tissue = annotation$Group,
                      MS_batch = annotation$MS_batch,
                      measurement_order = anno_lines(annotation$measurement_order),
                      col = list(Tissue = c("Tumor" = "red", "TANM" = "cadetblue"),
                                 MS_batch = colorRamp2(c(1, 4), c("green", "red"))))

heatmap <- Heatmap(oridata2, name = "∆ log2 LFQ", column_title = "Samples", row_title = "Proteins", 
        show_row_names = FALSE, show_column_names = FALSE, column_names_gp = gpar(fontsize = 5),
        show_row_dend = FALSE,
        cluster_columns = my_hclust_t,
        bottom_annotation = ha,
        heatmap_legend_param = list(title = "∆ log2 LFQ", title_position = "lefttop-rot"))
heatmap

tiff("results_2_humancohort_unsupervisedStats/Heatmap.tiff", units = "in", width = 6, height = 8, res = 300, pointsize =(8*100/72))
heatmap
dev.off()
```

# Immunemap
Including custom-made list of Immune Cell Markers
```{r}
oridata_immune <- column_to_rownames(oridata1, var = "Protein") %>%
  filter(rowSums(!is.na(.)) / ncol(.) > 0.3) %>%
  dplyr::select(annotation1$Sample)
subtractor <- apply(oridata_immune, 1, median, na.rm = TRUE)
oridata_immune <- oridata_immune - subtractor
oridata_immune <- rownames_to_column(oridata_immune, var = "UniProt_ID_Human")

immunelist <- read_excel("Data/CDMarkers_Profiling_ImmuneCells_Human_Mouse.xlsx") 
immunelist1 <- filter(immunelist, !is.na(`Key Marker`))

oridata_immune1 <- filter(oridata_immune, 
                          UniProt_ID_Human %in% immunelist1$UniProt_ID_Human) %>%
  left_join(dplyr::select(immunelist1, "UniProt_ID_Human", "Gene"), by = "UniProt_ID_Human") %>%
  arrange(Gene) %>%
  column_to_rownames(var = "Gene") %>%
  dplyr::select(-UniProt_ID_Human) %>%
  as.matrix() 

annotation_immune <- mutate(annotation, Group = factor(Group,
                                                            levels = c("Tumor", "TANM")))

immunelist_fit <- filter(immunelist1, Gene %in% rownames(oridata_immune1)) %>%
  arrange(Gene)

#CAVE! annotation file and plot matrix need to be in the exactly same order to match in the plot!
ha = columnAnnotation(Tissue = annotation_immune$Group,
                   col = list(Tissue = c("Tumor" = "red", "TANM" = "lightblue")), 
                   show_legend = FALSE)

hb = rowAnnotation(Category = immunelist_fit$`Key Marker`,
                   show_annotation_name = FALSE, 
                   show_legend = FALSE)

Heatmap <- Heatmap(oridata_immune1, name = "∆ log2 LFQ", column_title = " ", row_title = "", 
        col = colorRamp2(c(-5, 0, 5), c("blue", "grey93", "red")),
        show_row_names = TRUE, show_column_names = FALSE, 
        row_names_side = "left", row_names_gp = gpar(fontsize = 5),
        show_row_dend = FALSE, show_column_dend = FALSE, 
        cluster_rows = FALSE, cluster_columns = my_hclust_t,
        column_split = annotation_immune$cluster_K4, 
        row_split = immunelist_fit$`Key Marker`, row_title_rot = 0, 
        row_title_gp = gpar(fontsize = 5), row_title_side = "right",
        bottom_annotation = ha, right_annotation = hb, na_col = "white",
        heatmap_legend_param = list(title = "∆ log2 LFQ", title_position = "lefttop-rot"))
Heatmap

tiff("results_2_humancohort_unsupervisedStats/Heatmap_immune.tiff", units = "in", width = 6, height = 5, res = 300, pointsize =(6*100/72))
Heatmap
dev.off()
```

```{r}
immunelist2 <- filter(immunelist, !is.na(Gene))

oridata_immune2 <- filter(oridata_immune, 
                          UniProt_ID_Human %in% immunelist2$UniProt_ID_Human) %>%
  inner_join(dplyr::select(immunelist2, "UniProt_ID_Human", "Gene"), by = "UniProt_ID_Human") %>%
  unique() %>%
  remove_rownames() %>%
  column_to_rownames(var = "Gene") %>%
  dplyr::select(-UniProt_ID_Human) %>%
  as.matrix()

annotation_immune2 <- mutate(annotation, Group = factor(Group,
                                                            levels = c("Tumor", "TANM")))

immunelist_fit2 <- filter(immunelist2, Gene %in% rownames(oridata_immune2))
```

```{r}
#CAVE! annotation file and plot matrix need to be in the exactly same order to match in the plot!
ha = columnAnnotation(Tissue = annotation_immune$Group,
                   col = list(Tissue = c("Tumor" = "red", "TANM" = "lightblue")),
                   show_legend = FALSE)

hb = rowAnnotation(Category = immunelist_fit$`Key Marker`,
                   show_annotation_name = FALSE, 
                   show_legend = FALSE)

# perform clustering on imputed data, since NAs in original data do not permit clustering
oridata_immune_preimp <- filter(oridata, Protein %in% immunelist_fit2$UniProt_ID_Human) %>%
  column_to_rownames(var = "Protein") %>%
  t()
set.seed(546)
impfr_immune <- missForest(oridata_immune_preimp) # this takes ~ 3 minutes to compute

t_oridata_immune_imp <- impfr_immune$ximp
oridata_immune_imp <- t(t_oridata_immune_imp)
  
my_dist_t_immuneheatmap <- dist(oridata_immune_imp)
my_hclust_t_immuneheatmap <- hclust(my_dist_t_immuneheatmap)
plot(my_hclust_t_immuneheatmap)

# Heatmap
Heatmap <- Heatmap(oridata_immune2, name = "∆ log2 LFQ", column_title = " ", row_title = "", 
        col = colorRamp2(c(-5, 0, 5), c("blue", "grey93", "red")),
        show_row_names = TRUE, show_column_names = FALSE, 
        row_names_side = "left", row_names_gp = gpar(fontsize = 5),
        show_row_dend = FALSE, show_column_dend = FALSE, 
        cluster_rows = my_hclust_t_immuneheatmap, cluster_columns = FALSE,
        column_split = annotation_immune$Group, 
        #row_split = immunelist_fit$`Key Marker`, 
        row_title_rot = 0, 
        row_title_gp = gpar(fontsize = 5), row_title_side = "right",
        bottom_annotation = ha,  
        #right_annotation = hb,
        na_col = "white",
        heatmap_legend_param = list(title = "∆ log2 LFQ", title_position = "lefttop-rot"))
Heatmap

tiff("results_2_humancohort_unsupervisedStats/Heatmap_immune2.tiff", units = "in", width = 10, height = 10, res = 300, pointsize =(10*100/72))
Heatmap
dev.off()
```
